Renamed symbtrdataextractor/metadata.py -> symbtrdataextractor/MetadataExtractor.py
