+++ symbtrdataextractor/MetadataExtractor.py
